Buff unused chickens, buff bombs, nerf Viper
Slight adjustments to chickens to make more choices viable. Minor nerf to Viper metal cost to reduce the power of oneshot wave spawns which instantly win the game. Buffed bombs to be more viable again. They are still lower than their original aoe with the exception of Skuttle, but Skuttle isnt able to jump so a slight buff seems ok.